+++ CHANGELOG.md
@@ -6,6 +6,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+Unreleased
+----------
+
+### Added
+
+* Resourcelinks, accessible with the `getRaw()` method
+* Scenes, accessible with the `getRaw()` method
+* Possibility to activate a scene for a room
+
 v1.2.0 (2019-07-19)
 ----------