+++ CHANGELOG.md
@@ -11,11 +11,11 @@ Unreleased
 
 ### Added
 
+* Bridge discovery: possibility to find any Bridges in the current LAN. UPnP and N-UPnP methods supported.
 * Resourcelinks, accessible with the `getRaw()` method
 * Scenes, accessible with the `getRaw()` method
 * Possibility to activate a scene for a room
 * Other types of light groups besides just lights and zones are now supported too
-* N-UPnP Bridge Discovery Protocol: queries the Philips Hue portal for Hue Bridges in this LAN
 
 v1.2.0 (2019-07-19)
 ----------